+++ packages/feeds-client/src/feed/event-handlers/activity/handle-activity-added.test.ts
@@ -70,17 +70,24 @@ describe(handleActivityAdded.name, () => {
     const existing = generateActivityResponse();
     feed.state.partialNext({ activities: [existing] });
     const event = generateActivityAddedEvent();
+    const newActivitiesAddedSpy = vi.spyOn(feed, 'newActivitiesAdded' as any);
     handleActivityAdded.call(feed, event);
 
     const stateAfter = feed.currentState;
     expect(stateAfter.activities).toHaveLength(2);
     expect(stateAfter.activities?.[0]).toBe(existing);
     expect(stateAfter.activities?.[1]).toBe(event.activity);
+    expect(newActivitiesAddedSpy).toHaveBeenCalledWith([event.activity], {
+      fromWebSocket: true,
+    });
+
+    vi.resetAllMocks();
   });
 
   it('does not duplicate if activity already exists', () => {
     const existing = generateActivityResponse();
     feed.state.partialNext({ activities: [existing] });
+    const newActivitiesAddedSpy = vi.spyOn(feed, 'newActivitiesAdded' as any);
 
     const event = generateActivityAddedEvent({
       activity: { id: existing.id },
@@ -93,6 +100,9 @@ describe(handleActivityAdded.name, () => {
     expect(stateAfter).toBe(stateBefore);
     expect(stateAfter.activities).toHaveLength(1);
     expect(stateAfter.activities?.[0]).toBe(existing);
+    expect(newActivitiesAddedSpy).not.toHaveBeenCalled();
+
+    vi.resetAllMocks();
   });
 
   it(`onActivityAdded filters out activity if it returns false`, () => {