+++ packages/feeds-client/src/feeds-client/feeds-client.test.ts
@@ -1,6 +1,12 @@
-import { beforeEach, describe, expect, it } from 'vitest';
+import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { FeedsClient } from './feeds-client';
-import { generateActivityResponse } from '../test-utils';
+import {
+  generateActivityResponse,
+  generateFeedMemberResponse,
+  generateFeedResponse,
+  generateFollowResponse,
+} from '../test-utils';
+import { FeedOwnCapability } from '..';
 
 describe('Feeds client tests', () => {
   let client: FeedsClient;
@@ -27,5 +33,195 @@ describe('Feeds client tests', () => {
     expect(feeds).toContain(feedA);
     expect(feeds).toContain(feedC);
     expect(feeds).not.toContain(feedB);
+  });
+
+  describe('client.feed', () => {
+    it('should initialize feed state with the given data if feed does not exist', async () => {
+      const data = generateFeedResponse({ feed: 'timeline:feed' });
+      client['getOrCreateActiveFeed']({ group: 'timeline', id: 'feed', data });
+
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState,
+      ).toMatchObject(data);
+    });
+
+    it('should update feed state if data is newer than the current state', async () => {
+      const data = generateFeedResponse({
+        feed: 'timeline:feed',
+        follower_count: 5,
+      });
+      client['getOrCreateActiveFeed']({ group: 'timeline', id: 'feed', data });
+
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState,
+      ).toMatchObject(data);
+
+      const newData = { ...data };
+      newData.updated_at = new Date(data.updated_at.getTime() + 1000);
+      newData.follower_count = 10;
+      client['getOrCreateActiveFeed']({
+        group: 'timeline',
+        id: 'feed',
+        data: newData,
+      });
+
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState,
+      ).toMatchObject(newData);
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState.follower_count,
+      ).toBe(10);
+    });
+
+    it(`should not update feed state if data is older than the current state`, async () => {
+      const data = generateFeedResponse({
+        feed: 'timeline:feed',
+        follower_count: 5,
+      });
+      client['getOrCreateActiveFeed']({ group: 'timeline', id: 'feed', data });
+
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState,
+      ).toMatchObject(data);
+
+      const oldData = { ...data };
+      oldData.updated_at = new Date(data.updated_at.getTime() - 1000);
+      oldData.follower_count = 3;
+      client['getOrCreateActiveFeed']({
+        group: 'timeline',
+        id: 'feed',
+        data: oldData,
+      });
+
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState,
+      ).toMatchObject(data);
+      expect(
+        client['activeFeeds']['timeline:feed']?.currentState.follower_count,
+      ).toBe(5);
+    });
+
+    it(`should not update own_ fields if data is from WebSocket`, async () => {
+      const ownFollows = [
+        generateFollowResponse({
+          source_feed: generateFeedResponse({ feed: 'timeline:feed' }),
+          target_feed: generateFeedResponse({ feed: 'user:123' }),
+        }),
+      ];
+      const ownMembership = generateFeedMemberResponse();
+      const ownCapabilities = [FeedOwnCapability.ADD_ACTIVITY];
+      const data = generateFeedResponse({
+        feed: 'user:123',
+        follower_count: 5,
+        own_follows: ownFollows,
+        own_membership: ownMembership,
+        own_capabilities: ownCapabilities,
+      });
+      client['getOrCreateActiveFeed']({ group: 'user', id: '123', data });
+
+      const dataFromWebSocket = { ...data };
+      delete dataFromWebSocket.own_follows;
+      delete dataFromWebSocket.own_membership;
+      delete dataFromWebSocket.own_capabilities;
+      client['getOrCreateActiveFeed']({
+        group: 'user',
+        id: '123',
+        data: dataFromWebSocket,
+        fromWebSocket: true,
+      });
+
+      expect(client['activeFeeds']['user:123']?.currentState).toMatchObject(
+        data,
+      );
+      expect(
+        client['activeFeeds']['user:123']?.currentState.own_follows,
+      ).toEqual(ownFollows);
+      expect(
+        client['activeFeeds']['user:123']?.currentState.own_membership,
+      ).toEqual(ownMembership);
+      expect(
+        // @ts-expect-error - own_capabilities is currently excluded from type
+        client['activeFeeds']['user:123']?.currentState.own_capabilities,
+      ).toEqual(ownCapabilities);
+    });
+  });
+
+  it(`should only update _own fields which are changed`, async () => {
+    const ownFollows = [
+      generateFollowResponse({
+        source_feed: generateFeedResponse({ feed: 'timeline:feed' }),
+        target_feed: generateFeedResponse({ feed: 'user:123' }),
+      }),
+    ];
+    const ownMembership = generateFeedMemberResponse();
+    const ownCapabilities = [FeedOwnCapability.ADD_ACTIVITY];
+    const data = generateFeedResponse({
+      feed: 'user:123',
+      follower_count: 5,
+      own_follows: ownFollows,
+      own_membership: ownMembership,
+      own_capabilities: ownCapabilities,
+    });
+    client['getOrCreateActiveFeed']({ group: 'user', id: '123', data });
+
+    const spy = vi.fn();
+    const feed = client['activeFeeds']['user:123'];
+    feed.state.subscribe(spy);
+    spy.mockClear();
+
+    const newData = { ...data };
+    data.own_follows = [...ownFollows];
+    client['getOrCreateActiveFeed']({
+      group: 'user',
+      id: '123',
+      data: newData,
+      fromWebSocket: false,
+    });
+
+    expect(spy).toHaveBeenCalledTimes(0);
+
+    const newOwnFollows = [
+      ...ownFollows,
+      generateFollowResponse({
+        source_feed: generateFeedResponse({ feed: 'timeline:feed' }),
+        target_feed: generateFeedResponse({ feed: 'user:456' }),
+      }),
+    ];
+    newData.own_follows = newOwnFollows;
+
+    client['getOrCreateActiveFeed']({
+      group: 'user',
+      id: '123',
+      data: newData,
+    });
+
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy.mock.lastCall?.[0]).toMatchObject({
+      own_follows: newOwnFollows,
+    });
+
+    spy.mockClear();
+    newData.own_membership = { ...newData.own_membership! };
+    client['getOrCreateActiveFeed']({
+      group: 'user',
+      id: '123',
+      data: newData,
+    });
+
+    expect(spy).toHaveBeenCalledTimes(0);
+
+    newData.own_membership!.updated_at = new Date(
+      newData.own_membership!.updated_at.getTime() + 1000,
+    );
+    client['getOrCreateActiveFeed']({
+      group: 'user',
+      id: '123',
+      data: newData,
+    });
+
+    expect(spy).toHaveBeenCalledTimes(1);
+    expect(spy.mock.lastCall?.[0]).toMatchObject({
+      own_membership: newData.own_membership,
+    });
   });
 });